+++ .gitignore
@@ -1,1 +1,6 @@
 /bin/
+/target
+/.gradle/
+build/
+out/
+.idea/
