+++ src/main/java/com/apickledwalrus/papiaddon/Main.java
@@ -7,20 +7,12 @@ import java.io.IOException;
 
 public class Main extends JavaPlugin {
 	private static Main instance;
-	private static SkriptAddon addon;
 
-	public Main() {
-		if (instance == null) {
-			instance = this;
-		} else {
-			throw new IllegalStateException();
-		}
-	}
-
-	@Override
+  @Override
 	public void onEnable() {
+		instance = this;
 		try {
-			getAddonInstance().loadClasses("com.apickledwalrus.papiaddon.skript", "events", "expressions");
+			Skript.registerAddon(this).loadClasses("com.apickledwalrus.papiaddon.skript", "events", "expressions");
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
@@ -31,12 +23,5 @@ public class Main extends JavaPlugin {
 			throw new IllegalStateException();
 		}
 		return instance;
-	}
-
-	public static SkriptAddon getAddonInstance() {
-		if (addon == null) {
-			addon = Skript.registerAddon(getInstance());
-		}
-		return addon;
 	}
 }
